+++ lib/ayeaye/connectors/ndjson_connector.py
@@ -73,6 +73,9 @@ class NdjsonConnector(FileBasedConnector):
             self.approx_position += len(str(r))
             yield Pinnate(data=r)
 
+        # reduce the number of open file handles when the whole file has been read
+        self.close_connection()
+
     @property
     def data(self):
         raise NotImplementedError("TODO")