+++ lib/ayeaye/connectors/parquet_connector.py
@@ -102,6 +102,9 @@ class ParquetConnector(DataConnector):
                 yield Pinnate(data=row_as_dict)
                 self.current_row += 1
 
+        # reduce the number of open file handles when the whole file has been read
+        self.close_connection()
+
     @property
     def data(self):
         raise NotImplementedError("TODO")